installer for GMDX, HX, and Revision
